+++ gluecodium/src/test/resources/smoke/external_types/output/dart/lib/src/smoke/structs.dart
@@ -4,12 +4,16 @@ import 'package:library/src/_native_base.dart' as __lib;
 import 'package:library/src/_token_cache.dart' as __lib;
 import 'package:library/src/builtin_types__conversion.dart';
 import 'package:library/src/generic_types__conversion.dart';
+import 'package:meta/meta.dart';
 abstract class Structs {
   /// @nodoc
   @Deprecated("Does nothing")
   void release();
-  static Structs_ExternalStruct getExternalStruct() => Structs$Impl.getExternalStruct();
-  static Structs_AnotherExternalStruct getAnotherExternalStruct() => Structs$Impl.getAnotherExternalStruct();
+  static Structs_ExternalStruct getExternalStruct() => $prototype.getExternalStruct();
+  static Structs_AnotherExternalStruct getAnotherExternalStruct() => $prototype.getAnotherExternalStruct();
+  /// @nodoc
+  @visibleForTesting
+  static dynamic $prototype = Structs$Impl(Pointer<Void>.fromAddress(0));
 }
 class Structs_ExternalStruct {
   String stringField;
@@ -180,11 +184,13 @@ final _smokeStructsReleaseHandle = __lib.catchArgumentError(() => __lib.nativeLi
     Void Function(Pointer<Void>),
     void Function(Pointer<Void>)
   >('library_smoke_Structs_release_handle'));
+/// @nodoc
+@visibleForTesting
 class Structs$Impl extends __lib.NativeBase implements Structs {
   Structs$Impl(Pointer<Void> handle) : super(handle);
   @override
   void release() {}
-  static Structs_ExternalStruct getExternalStruct() {
+  Structs_ExternalStruct getExternalStruct() {
     final _getExternalStructFfi = __lib.catchArgumentError(() => __lib.nativeLibrary.lookupFunction<Pointer<Void> Function(Int32), Pointer<Void> Function(int)>('library_smoke_Structs_getExternalStruct'));
     final __resultHandle = _getExternalStructFfi(__lib.LibraryContext.isolateId);
     try {
@@ -193,7 +199,7 @@ class Structs$Impl extends __lib.NativeBase implements Structs {
       smokeStructsExternalstructReleaseFfiHandle(__resultHandle);
     }
   }
-  static Structs_AnotherExternalStruct getAnotherExternalStruct() {
+  Structs_AnotherExternalStruct getAnotherExternalStruct() {
     final _getAnotherExternalStructFfi = __lib.catchArgumentError(() => __lib.nativeLibrary.lookupFunction<Pointer<Void> Function(Int32), Pointer<Void> Function(int)>('library_smoke_Structs_getAnotherExternalStruct'));
     final __resultHandle = _getAnotherExternalStructFfi(__lib.LibraryContext.isolateId);
     try {