+++ gluecodium/src/test/resources/smoke/lambdas/output/dart/lib/smoke.dart
@@ -1,5 +1,5 @@
-export 'src/smoke/class_with_internal_lambda.dart' show ClassWithInternalLambda;
-export 'src/smoke/lambdas.dart' show Lambdas, Lambdas_Confuser, Lambdas_Consumer, Lambdas_Indexer, Lambdas_NullableConfuser, Lambdas_Producer;
+export 'src/smoke/class_with_internal_lambda.dart' show ClassWithInternalLambda, ClassWithInternalLambda$Impl;
+export 'src/smoke/lambdas.dart' show Lambdas, Lambdas$Impl, Lambdas_Confuser, Lambdas_Consumer, Lambdas_Indexer, Lambdas_NullableConfuser, Lambdas_Producer;
 export 'src/smoke/lambdas_declaration_order.dart' show LambdasDeclarationOrder, LambdasDeclarationOrder_SomeCallback, LambdasDeclarationOrder_SomeStruct;
 export 'src/smoke/lambdas_interface.dart' show LambdasInterface, LambdasInterface_TakeScreenshotCallback;
 export 'src/smoke/lambdas_with_structured_types.dart' show LambdasWithStructuredTypes, LambdasWithStructuredTypes_ClassCallback, LambdasWithStructuredTypes_StructCallback;